+++ src/jquery.fancytree.glyph.js
@@ -47,7 +47,7 @@ $.ui.fancytree.registerExtension({
 			dragHelper: "icon-caret-right",
 			dropMarker: "icon-caret-right",
 			// Default node icons.
-			// (Use tree.options.iconClass callback to define custom icons
+			// (Use tree.options.icon callback to define custom icons
 			// based on node data)
 			doc: "icon-file-alt",
 			docOpen: "icon-file-alt",
@@ -103,8 +103,8 @@ $.ui.fancytree.registerExtension({
 			span.className = "fancytree-checkbox " + map[icon];
 		}
 
-		// Icon (note that this does not match .fancytree-custom-icon, that might
-		// be set by opts.iconClass)
+		// Standard icon (note that this does not match .fancytree-custom-icon, 
+		// that might be set by opts.icon callbacks)
 		span = $span.children("span.fancytree-icon").get(0);
 		if( span ){
 			if( node.folder ){