+++ test/dnd-sortable/test-dnd-sortable-5.js
@@ -27,7 +27,7 @@ $(document).ready(function(){
 				 {title: "Lazy1", folder: true, lazy: true }
 				 ],
 //		checkbox: true,
-//		icons: false,
+//		icon: false,
 //        activeVisible: true,
 		dnd: {
 			preventVoidMoves: true, // Prevent dropping nodes 'before self', etc.
@@ -119,7 +119,7 @@ $(document).ready(function(){
 	$("#fancytree2").fancytree({
 		extensions: ["dnd"],
 		selectMode: 2, // multi-select
-		icons: false,
+		icon: false,
 //                source: "#sourceItems",
 		dnd: {
 			preventVoidMoves: true, // Prevent dropping nodes 'before self', etc.